refactor: move `validate_field` to `FieldDefinition` impl
